* Update crowdin.yml

* updated dependencies

* Better tv navigation (#281)

* updated navigation for authentication fragment

* added tv banner

close #276

* fix wrong library version

* avoid crash when clicking urls on devices without a browser

* Update fragment_authentication.xml

* Update fragment_authentication.xml

* Update fragment_user_profile.xml

* Update fragment_folder_list.xml

* Update fragment_tab_lists.xml

* removed focusable from cards

* updated search fragment navigation

* updated folder fragment navigation

* updated new download fragment

* Update TorrentFilePickerFragment.kt

* removed unneeded attributes from buttons

* updated kodi list layout

* Update dialog_kodi_management.xml

* replaced imagebutton

* Update crowdin.yml

* Update Crowdin configuration file

* updated translations

* linting

* automatic torrent file selection if only a single file is available

* Add plugins repository support (#285)

close #278 


* updated dependencies

* added new fragment for repository management

* added search bar functionality

* download plugin into own repo folder

* addded function to retrieve installed plugins paired with repositories

* added plugin uninstall

* added install all from repo

* updated remove all plugins and plugin settings

* moved logout under settings

replaced logout button with settings for android tv users

* added function to remove repository plugins

* added uninstall repository option

* added readme for plugins

* updated theme

see #284

* added index property to table parser

close #201

* avoid removing the default repository from the database

* bumped version for release

* formatted code

* formatting

* formatting

* Update item_plugin_repository_repo.xml

* updated locally installed plugins management

* removed unused search plugins assets folder

* added add repository dialog

* removed references to old plugins pack

* fixed add repository dialog

* added more categories to plugins

close #289

* better new repository hadnling

* bumped version for release

* formatted code

* removed plugins step from CI

* fixed kodi dialogs

* added check for exception

* fix import repository crash

adding wrong links crashed the app

* fix shield

* updated room

* updated dependencies

* bumped version for release

* restored correct library version

* updated dependencies

* Create dependency updates.run.xml

* added option to disable downloads vibration

close #294

* fixed failing web link regex

* Feat/select kodi (#298)

* Added option to pick kodi server when sending links

close #246

* possible fix for crashes

* fixed possible crash

* bumped version for release

* replaced ktlint with ktfmt

* formatted all the code

used new formatter

* replace lint checker in CI

* removed lint report upload

not generated by new plugin

+++ app/app/src/main/java/com/github/livingwithhippos/unchained/data/repository/PluginRepository.kt
@@ -5,32 +5,33 @@ import android.net.Uri
 import com.github.livingwithhippos.unchained.plugins.model.Plugin
 import com.github.livingwithhippos.unchained.repository.model.RepositoryListItem
 import com.github.livingwithhippos.unchained.utilities.MANUAL_PLUGINS_REPOSITORY_NAME
-import com.github.livingwithhippos.unchained.utilities.getPluginFilename
 import com.github.livingwithhippos.unchained.utilities.getManualPluginFilename
+import com.github.livingwithhippos.unchained.utilities.getPluginFilename
 import com.github.livingwithhippos.unchained.utilities.getRepositoryString
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import timber.log.Timber
 import java.io.File
 import java.io.IOException
 import javax.inject.Inject
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import timber.log.Timber
 
 class PluginRepository @Inject constructor() {
 
     // todo: inject
-    private val pluginAdapter: JsonAdapter<Plugin> = Moshi.Builder()
-        .build()
-        .adapter(Plugin::class.java)
+    private val pluginAdapter: JsonAdapter<Plugin> =
+        Moshi.Builder().build().adapter(Plugin::class.java)
 
-    suspend fun getPlugins(context: Context, manuallyInstalledOnly: Boolean = false): Pair<List<Plugin>, Int> =
+    suspend fun getPlugins(
+        context: Context,
+        manuallyInstalledOnly: Boolean = false
+    ): Pair<List<Plugin>, Int> =
         withContext(Dispatchers.IO) {
-
             val pluginFiles = mutableListOf<File>()
             val pluginFolder = context.getDir("plugins", Context.MODE_PRIVATE)
             if (pluginFolder.exists()) {
-                if(manuallyInstalledOnly) {
+                if (manuallyInstalledOnly) {
                     val localPluginFolder = File(pluginFolder, MANUAL_PLUGINS_REPOSITORY_NAME)
                     if (localPluginFolder.exists()) {
                         pluginFolder.walk().forEach {
@@ -56,10 +57,7 @@ class PluginRepository @Inject constructor() {
                 try {
                     val json = file.readText()
                     val plugin: Plugin? = pluginAdapter.fromJson(json)
-                    if (plugin != null)
-                        plugins.add(plugin)
-                    else
-                        errors++
+                    if (plugin != null) plugins.add(plugin) else errors++
                 } catch (ex: Exception) {
                     Timber.e("Exception while parsing json plugin: $ex")
                 }
@@ -71,11 +69,11 @@ class PluginRepository @Inject constructor() {
     suspend fun getPluginsWithFolders(context: Context): LocalPlugins =
         withContext(Dispatchers.IO) {
 
-            /**
-             * list of plugin files associated with the repository folder name
-             */
+            /** list of plugin files associated with the repository folder name */
 
-            // note: the actual folder name may be different from the one put here. For example it could get changed to "app_plugins"
+            // note: the actual folder name may be different from the one put here. For example it
+            // could
+            // get changed to "app_plugins"
             // it is still deterministic
             val pluginFolder = context.getDir("plugins", Context.MODE_PRIVATE)
             val pluginRepoFileAssociation = mutableMapOf<String, List<File>>()
@@ -94,10 +92,9 @@ class PluginRepository @Inject constructor() {
                             }
                             repoFolder = currentFile.name
                             Timber.d("Found folder $repoFolder")
-                        } else if (currentFile.isFile && currentFile.name.endsWith(
-                                TYPE_UNCHAINED,
-                                ignoreCase = true
-                            )
+                        } else if (
+                            currentFile.isFile &&
+                                currentFile.name.endsWith(TYPE_UNCHAINED, ignoreCase = true)
                         ) {
                             Timber.d("Found plugin ${currentFile.name}")
                             files.add(currentFile)
@@ -127,10 +124,7 @@ class PluginRepository @Inject constructor() {
                     try {
                         val json = pluginFile.readText()
                         val plugin: Plugin? = pluginAdapter.fromJson(json)
-                        if (plugin != null)
-                            repoList.add(plugin)
-                        else
-                            errors++
+                        if (plugin != null) repoList.add(plugin) else errors++
                     } catch (ex: Exception) {
                         Timber.e("Exception while parsing json plugin: $ex")
                         errors++
@@ -142,11 +136,19 @@ class PluginRepository @Inject constructor() {
             LocalPlugins(pluginsData, errors)
         }
 
-    suspend fun removePlugin(context: Context, repository: String, author: String?, name: String): Boolean =
+    suspend fun removePlugin(
+        context: Context,
+        repository: String,
+        author: String?,
+        name: String
+    ): Boolean =
         withContext(Dispatchers.IO) {
             val pluginFolder = context.getDir("plugins", Context.MODE_PRIVATE)
             val repoName = getRepositoryString(repository)
-            val filename = if (repository == MANUAL_PLUGINS_REPOSITORY_NAME) getManualPluginFilename(author, name) else getPluginFilename(name)
+            val filename =
+                if (repository == MANUAL_PLUGINS_REPOSITORY_NAME)
+                    getManualPluginFilename(author, name)
+                else getPluginFilename(name)
 
             if (!pluginFolder.exists()) {
                 Timber.e("Plugin folder not found")
@@ -174,9 +176,11 @@ class PluginRepository @Inject constructor() {
             }
         }
 
-    suspend fun removePlugin(context: Context, repository: String, plugin: Plugin): Boolean = removePlugin(context, repository, plugin.author, plugin.name)
+    suspend fun removePlugin(context: Context, repository: String, plugin: Plugin): Boolean =
+        removePlugin(context, repository, plugin.author, plugin.name)
 
-    suspend fun removePlugin(context: Context, plugin: RepositoryListItem.Plugin): Boolean = removePlugin(context, plugin.repository, plugin.author, plugin.name)
+    suspend fun removePlugin(context: Context, plugin: RepositoryListItem.Plugin): Boolean =
+        removePlugin(context, plugin.repository, plugin.author, plugin.name)
 
     private fun getPluginFromJSON(json: String): Plugin? {
         return try {
@@ -195,12 +199,13 @@ class PluginRepository @Inject constructor() {
             pluginFolder.walk().forEach {
                 if (it.isFile) {
                     pluginCounter++
-                    if (it.delete())
-                        deleteCounter++
+                    if (it.delete()) deleteCounter++
                 }
             }
             val deleted = pluginFolder.deleteRecursively()
-            Timber.d("Found $pluginCounter plugins, deleted $deleteCounter plugins. Main directory deleted: $deleted")
+            Timber.d(
+                "Found $pluginCounter plugins, deleted $deleteCounter plugins. Main directory deleted: $deleted"
+            )
             deleteCounter
         } catch (e: SecurityException) {
             Timber.d("Security exception deleting plugins files: ${e.message}")
@@ -216,12 +221,8 @@ class PluginRepository @Inject constructor() {
             if (repoFolder.exists()) {
                 val pluginCounter = repoFolder.listFiles()?.size ?: 0
                 val deleted = repoFolder.deleteRecursively()
-                if (deleted)
-                    return pluginCounter
-                else
-                    return 0
-            } else
-                Timber.d("No plugin folder for repository $repositoryUrl found")
+                if (deleted) return pluginCounter else return 0
+            } else Timber.d("No plugin folder for repository $repositoryUrl found")
             0
         } catch (e: SecurityException) {
             Timber.e("Security exception deleting plugins files: ${e.message}")
@@ -229,31 +230,25 @@ class PluginRepository @Inject constructor() {
         }
     }
 
-    suspend fun addExternalPlugin(
-        pluginsFolder: File,
-        pluginFile: File
-    ): Boolean = withContext(Dispatchers.IO) {
-        try {
-            val savedPluginPath = File(pluginsFolder, pluginFile.name)
-            // remove file if already existing
-            if (savedPluginPath.exists())
-                savedPluginPath.delete()
-            // copy plugin from cache to internal memory
-            pluginFile.inputStream().use { inputStream ->
-                savedPluginPath.outputStream().use {
-                    inputStream.copyTo(it)
+    suspend fun addExternalPlugin(pluginsFolder: File, pluginFile: File): Boolean =
+        withContext(Dispatchers.IO) {
+            try {
+                val savedPluginPath = File(pluginsFolder, pluginFile.name)
+                // remove file if already existing
+                if (savedPluginPath.exists()) savedPluginPath.delete()
+                // copy plugin from cache to internal memory
+                pluginFile.inputStream().use { inputStream ->
+                    savedPluginPath.outputStream().use { inputStream.copyTo(it) }
                 }
+            } catch (exception: Exception) {
+                Timber.e("Error saving the plugin ${pluginFile.name}: ${exception.message}")
+                return@withContext false
             }
-        } catch (exception: Exception) {
-            Timber.e("Error saving the plugin ${pluginFile.name}: ${exception.message}")
-            return@withContext false
+            return@withContext true
         }
-        return@withContext true
-    }
 
     suspend fun readPassedPlugin(context: Context, data: Uri): Plugin? =
         withContext(Dispatchers.IO) {
-
             val filename = data.path?.split("/")?.last()
             if (filename != null) {
                 try {
@@ -287,64 +282,56 @@ class PluginRepository @Inject constructor() {
         context: Context,
         plugin: Plugin,
         repositoryURL: String?
-    ): InstallResult = withContext(Dispatchers.IO) {
-        // we use the repo link hash as folder name for all the plugins from that repo
-        val repoName = if (repositoryURL != null)
-            getRepositoryString(repositoryURL)
-        else
-            MANUAL_PLUGINS_REPOSITORY_NAME
+    ): InstallResult =
+        withContext(Dispatchers.IO) {
+            // we use the repo link hash as folder name for all the plugins from that repo
+            val repoName =
+                if (repositoryURL != null) getRepositoryString(repositoryURL)
+                else MANUAL_PLUGINS_REPOSITORY_NAME
 
-        /*
-         issue: we have 3 possible installation modes:
-         1. local file shared to the app
-         2. file url shared to the app
-         3. download from a repository
+            /*
+            issue: we have 3 possible installation modes:
+            1. local file shared to the app
+            2. file url shared to the app
+            3. download from a repository
 
-         while 3 is easy to manage, 1 and 2 means the file will go to a custom folder
-         how to manage deletion/update of these files?
-         In particular let's say we have 2 plugins with the same name, how to differentiate them?
-         */
+            while 3 is easy to manage, 1 and 2 means the file will go to a custom folder
+            how to manage deletion/update of these files?
+            In particular let's say we have 2 plugins with the same name, how to differentiate them?
+            */
 
-        val filename = if (repositoryURL != null)
-            getPluginFilename(plugin.name)
-        else
-            getManualPluginFilename(plugin.author, plugin.name)
+            val filename =
+                if (repositoryURL != null) getPluginFilename(plugin.name)
+                else getManualPluginFilename(plugin.author, plugin.name)
 
-        try {
-            val pluginFolder = context.getDir("plugins", Context.MODE_PRIVATE)
-            if (!pluginFolder.exists())
-                pluginFolder.mkdirs()
+            try {
+                val pluginFolder = context.getDir("plugins", Context.MODE_PRIVATE)
+                if (!pluginFolder.exists()) pluginFolder.mkdirs()
 
-            val repoFolder = File(pluginFolder, repoName)
-            if (!repoFolder.exists())
-                repoFolder.mkdirs()
+                val repoFolder = File(pluginFolder, repoName)
+                if (!repoFolder.exists()) repoFolder.mkdirs()
 
-            val file = File(repoFolder, filename)
+                val file = File(repoFolder, filename)
 
-            // this will overwrite the file, no deletion needed
-            // utf 8 is enough?
-            file.writeText(pluginAdapter.toJson(plugin))
-        } catch (exception: IOException) {
-            Timber.e("Error saving into memory the plugin $filename: ${exception.message}")
-            return@withContext InstallResult.Error(exception)
-        } catch (exception: Exception) {
-            Timber.e("Unknown error adding the plugin $filename: ${exception.message}")
-            return@withContext InstallResult.Error(exception)
+                // this will overwrite the file, no deletion needed
+                // utf 8 is enough?
+                file.writeText(pluginAdapter.toJson(plugin))
+            } catch (exception: IOException) {
+                Timber.e("Error saving into memory the plugin $filename: ${exception.message}")
+                return@withContext InstallResult.Error(exception)
+            } catch (exception: Exception) {
+                Timber.e("Unknown error adding the plugin $filename: ${exception.message}")
+                return@withContext InstallResult.Error(exception)
+            }
+
+            return@withContext InstallResult.Installed
         }
-
-        return@withContext InstallResult.Installed
-    }
 
     suspend fun savePluginFromDisk(context: Context, data: Uri): InstallResult {
         val plugin: Plugin? = readPassedPlugin(context, data)
         if (plugin != null) {
-            if (!plugin.isCompatible())
-                return InstallResult.Incompatible
-            val saved = savePlugin(
-                context,
-                plugin,
-                null
-            )
+            if (!plugin.isCompatible()) return InstallResult.Incompatible
+            val saved = savePlugin(context, plugin, null)
             return saved
         }
         return InstallResult.Error(IllegalArgumentException("Could not load plugin from $data"))
@@ -361,7 +348,4 @@ sealed class InstallResult {
     object Incompatible : InstallResult()
 }
 
-data class LocalPlugins(
-    val pluginsData: Map<String, List<Plugin>>,
-    val errors: Int
-)
+data class LocalPlugins(val pluginsData: Map<String, List<Plugin>>, val errors: Int)
